+++ src/types.ts
@@ -4,6 +4,11 @@ export interface SaveMessage {
 	data: string,
 }
 
+export interface AutosaveRequest {
+	type: 'autosave',
+	data: string,
+}
+
 // Show the "discard changes" button
 export interface ShowCloseButtonRequest {
 	type: 'showCloseUnsavedBtn',
@@ -21,4 +26,5 @@ export interface InitialSvgDataRequest {
 	type: 'getInitialData',
 }
 
-export type WebViewMessage = SaveMessage | ShowCloseButtonRequest | HideCloseButtonRequest | ResumeEditingMessage | InitialSvgDataRequest;
+export type WebViewMessage =
+	SaveMessage | ShowCloseButtonRequest | HideCloseButtonRequest | ResumeEditingMessage | InitialSvgDataRequest | AutosaveRequest;